start Task 2
no completed

+++ src/Point.java
@@ -1,25 +1,23 @@
-import org.sparkproject.dmg.pmml.False;
-
 public class Point {
-    private final float x;
-    private final float y;
+    private final double x;
+    private final double y;
 
-    public Point(float x, float y){
+    public Point(double x, double y){
         this.x = x;
         this.y = y;
     }
 
-    public float getX(){
+    public double getX(){
         return this.x;
     }
 
-    public float getY(){
+    public double getY(){
         return this.y;
     }
 
     public double distanceTo(Point other) {
-        float deltaX = other.getX() - this.getX();
-        float deltaY = other.getY() - this.getY();
+        double deltaX = other.getX() - this.getX();
+        double deltaY = other.getY() - this.getY();
         // return Math.sqrt(deltaX * deltaX + deltaY * deltaY);
         return Math.sqrt(Math.pow(deltaX, 2) + Math.pow(deltaY, 2));
     }
